Mass Change of Object Structures for Ease

+++ ArcanaRush/Objects/Bullet.h
@@ -5,9 +5,9 @@
 class Bullet : public Entity {
 
 public:
-	float damage;
+	int damage;
 
-	Bullet(float x, float y, float w, float h, float d, float v, int mv, float dmg);
+	Bullet(float x, float y, float w, float h, float d, float v, int mv, int dmg);
 
 };
