+++ ArcanaRush/Objects/Game.h
@@ -9,7 +9,7 @@
 # include "Boss.h"
 # include "Bullet.h"
 
-const float PI = 3.14159265359;
+const double PI = (float)3.14159265359;
 
 class Game{
 	float score;
